perf(diagnostics): adopt u32 SourceSpan from oxc-miette
oxc-miette now stores span offsets/lengths as u32 (shrinking LabeledSpan
from 48 to 40 bytes). Drop the usize<->u32 round-trip casts at every
label construction and read site now that the API is u32-native.

+++ apps/oxlint/src/lsp/error_with_position.rs
@@ -76,9 +76,6 @@ fn miette_severity_to_lsp_severity(value: Severity) -> DiagnosticSeverity {
         Severity::Advice => DiagnosticSeverity::HINT,
     }
 }
-// clippy: the source field is checked and assumed to be less than 4GB, and
-// we assume that the fix offset will not exceed 2GB in either direction
-#[expect(clippy::cast_possible_truncation)]
 pub fn message_to_lsp_diagnostic(
     message: Message,
     uri: &Uri,
@@ -101,10 +98,10 @@ pub fn message_to_lsp_diagnostic(
         spans
             .iter()
             .map(|span| {
-                let offset = span.offset() as u32;
+                let offset = span.offset();
                 let start_position = offset_to_position(rope, offset, source_text);
                 let end_position =
-                    offset_to_position(rope, offset + span.len() as u32, source_text);
+                    offset_to_position(rope, offset + span.len(), source_text);
 
                 ls_types::DiagnosticRelatedInformation {
                     location: ls_types::Location {
