+++ crates/oxc_linter/src/fixer/mod.rs
@@ -273,13 +273,12 @@ pub struct Message {
 }
 
 impl Message {
-    #[expect(clippy::cast_possible_truncation)] // for `as u32`
     pub fn new(error: OxcDiagnostic, fixes: PossibleFixes) -> Self {
         let span = error
             .labels
             .as_ref()
             .and_then(|labels| labels.iter().find(|span| span.primary()).or_else(|| labels.first()))
-            .map(|span| Span::new(span.offset() as u32, (span.offset() + span.len()) as u32))
+            .map(|span| Span::new(span.offset(), span.offset() + span.len()))
             .unwrap_or_default();
 
         Self { error, span, fixes, fixed: false, section_offset: 0, rule: None }
@@ -305,7 +304,7 @@ impl Message {
 
         if let Some(labels) = &mut self.error.labels {
             for label in labels {
-                label.set_span_offset(label.offset().saturating_add(offset as usize));
+                label.set_span_offset(label.offset().saturating_add(offset));
             }
         }
